Add files via upload
The project aims to enhance the security of communication channels by implementing a robust 
cryptographic framework. Authenticated key exchange protocols play an important role in securing 
communication and are extensively deployed for use in various real-world network applications. DiffieHellman key exchange protocol has been widely used for establishing a secret key in two-party 
communications without using any third-party key management entity. But, Diffie-Hellman key 
exchange protocol does not authenticate the users in the communication and is vulnerable to man in-themiddle attack when the environment is wireless 
The Diffie-Hellman algorithm enables the secure exchange of a shared secret key over an insecure 
channel, forming the foundation of the system. AES, a robust encryption standard, ensures the 
confidentiality of the transmitted data by encrypting it using the derived shared secret key. 
To safeguard data integrity, SHA hashing is employed to generate a unique digital fingerprint of the 
message. By comparing the hash of the original message with the hash of the decrypted message, any 
unauthorized modifications during transmission can be detected. The integration of these cryptographic 
techniques provides a comprehensive solution for secure communication, protecting both the 
confidentiality and integrity of sensitive information